requirements.txt
asgiref==3.4.1
click==8.0.1
fastapi==0.68.0
gunicorn==20.1.0
h11==0.12.0
mysql-connector-python==8.0.23
protobuf==3.17.3
pydantic==1.8.2
python-dotenv==0.19.0
python-multipart==0.0.5
six==1.16.0
starlette==0.14.2
typing-extensions==3.10.0.0
uvicorn==0.14.0
zope.event==4.5.0
zope.interface==5.4.0
